buildah/connection add support of specific user
Add possibility to use a specific user, either defined by --user argument on
command line, or ansible_user, to connect to containers.

It is inspired from
https://github.com/containers/ansible-podman-collections/pull/19 the equivalent
for podman connection.

It was laso required to change the method to put file, from a mount mechanism,
to the buildah copy function, to allow to set correctly the permissions of
pushed files.

Fixes:  containers/ansible-podman-collections#25