Switch to use .get() to prevent keyerror
When trying to leverage the PodmanManager from another module, we can
end up with some parameters not necessarily defined (e.g. no defaults).
In many places in the code we have self.module_params[key] is None or a
variation of that.  If the key is not set, this results in a key error
but we can avoid this simply by using .get() instead. This change
updates these conditionals to prevent key errors. This change also
switches up how the default variables are done by starting with module
parameters and only switching to default if the value is none or it's
not defined.