fix check for read-only change of root image
The --read-only option is not checked for changes between runs of
ansible thus the container is not recreated if the user changes
the root image between read-write to read-only.

Default for the read_only variable set to False in line with
documentation and so that later checks are simplified.

This is a fix for bug #383.

Signed-off-by: Andrew <rubiksdot@grue.cc>