Fix log_level module parameter
The `log_level` module parameter is not guaranteed by Ansible and not
something we have a default for within the library. If the `log_level`
parameter is not defined the following error will be raised resulting
in deployment failure [0]. This change adjusts how the `log_level`
parameter is collected so that we can ensure the module isn't failing
unnecessarily.

[0] https://paste.opendev.org/show/b3dJmgeuZKna74RQo5vy/

Signed-off-by: Kevin Carter <kecarter@redhat.com>