check --mount, --tmpfs, --volume and VOLUME for changes
Combine diffparam_mount(), diffparam_tmpfs() and diffparam_volume()
and add handling of VOLUME into one function that allows for movement
of definitions between the various ways of specifying mount points
as long as the options are equivalent.

This necessitated translations of the various ways of defining
mount points into one single "language" so that the definitions
may be compared and checked for similarity. prep_mount_args_for_comp()
does this work and includes within it the default options for the
various CLI arguments and VOLUME.

The consequence of this is that the diff returned by the function
no longer literally represents the real-life arguments passed but,
rather, the translations of those arguments.

Signed-off-by: Andrew <rubiksdot@grue.cc>