fix pod running status for older podman versions
Older podman versions, such as podman 1.6.4 on CentOS 7, do not have the
pod status in the 'podman pod inspect' output. Added an attribute and a
method to PodmanPod to fetch the 'podman pod ps' output. Added an
additional status check to the exising running property that uses the
'ps' info.

Fixes #499

Signed-off-by: antonc42 <antonc42@users.noreply.github.com>